Implement logic for generating enum classes
Content for enum classes are retrieved from ValuSet and CodeSet
Terminology resources. Some enums are generated as nested classes
in the generated FHIR models, if the element declared in the
StructureDefinition is not a common binding element. Example is
the NameUse enum class of class HumanName (used in Patient model).

Signed-off-by: Elly Kitoto <junkmailstoelly@gmail.com>